backend/requirements.txt
fastapi
uvicorn
GROQ
pydantic
python-dotenv
